+++ TextAdventureServer/src/HelperObjects/Hitbox.java
@@ -2,16 +2,16 @@ package HelperObjects;
 
 public class Hitbox {
 	private boolean isSquare;
-	private int radius; // halbe seitenlänge bei square
+	private double radius; // halbe seitenlänge bei square
 
-	public Hitbox(boolean isSquare, int radius) {
+	public Hitbox(boolean isSquare, double radius) {
 		this.radius = radius;
 		this.isSquare = isSquare;
 	}
 
 	public boolean isHit(Hitbox checkBox, double xOffset, double yOffset) {
-		int x = (int) Math.abs(xOffset);
-		int y = (int) Math.abs(yOffset);
+		double x = Math.abs(xOffset);
+		double y = Math.abs(yOffset);
 		if (isSquare) {
 			return squareHit(checkBox, x, y);
 		} else {
@@ -19,20 +19,20 @@ public class Hitbox {
 		}
 	}
 
-	private boolean squareHit(Hitbox checkBox, int xOffset, int yOffset) {
-		int radiusSum = radius + checkBox.getRadius();
+	private boolean squareHit(Hitbox checkBox, double xOffset, double yOffset) {
+		double radiusSum = radius + checkBox.getRadius();
 		if (checkBox.isSquare) {
 			return xOffset <= radiusSum && yOffset <= radiusSum;
 		} else {
-			int cradius = checkBox.getRadius();
+			double cradius = checkBox.getRadius();
 			return checkOverlap(cradius, 0, 0, xOffset - radius / 2, yOffset - radius / 2, xOffset + radius / 2,
 					yOffset + radius / 2);
 		}
 	}
 
-	private boolean circleHit(Hitbox checkBox, int xOffset, int yOffset) {
+	private boolean circleHit(Hitbox checkBox, double xOffset, double yOffset) {
 		if (checkBox.isSquare) {
-			int sradius = checkBox.getRadius();
+			double sradius = checkBox.getRadius();
 			return checkOverlap(radius, 0, 0, xOffset - sradius / 2, yOffset - sradius / 2, xOffset + sradius / 2,
 					yOffset + sradius / 2);
 		} else {
@@ -40,12 +40,12 @@ public class Hitbox {
 		}
 	}
 
-	private boolean checkOverlap(int R, int Xc, int Yc, int X1, int Y1, int X2, int Y2) {
-		int Xn = Math.max(X1, Math.min(Xc, X2));
-		int Yn = Math.max(Y1, Math.min(Yc, Y2));
+	private boolean checkOverlap(double R, double Xc, double Yc, double X1, double Y1, double X2, double Y2) {
+		double Xn = Math.max(X1, Math.min(Xc, X2));
+		double Yn = Math.max(Y1, Math.min(Yc, Y2));
 
-		int Dx = Xn - Xc;
-		int Dy = Yn - Yc;
+		double Dx = Xn - Xc;
+		double Dy = Yn - Yc;
 		return (Dx * Dx + Dy * Dy) <= R * R;
 	}
 
@@ -53,7 +53,7 @@ public class Hitbox {
 		return isSquare;
 	}
 
-	public int getRadius() {
+	public double getRadius() {
 		return radius;
 	}
